+++ tests/Unit/SharePointDsc/SharePointDsc.SPFarm.Tests.ps1
@@ -33,7 +33,7 @@ function Invoke-TestSetup
 
     $script:testEnvironment = Initialize-TestEnvironment `
         -DSCModuleName $script:DSCModuleName `
-        -DSCResourceName $script:DSCResourceFullName `
+        -DscResourceName $script:DSCResourceFullName `
         -ResourceType 'Mof' `
         -TestType 'Unit'
 }
@@ -108,6 +108,31 @@ try
                         DeveloperDashboardSettings = $developerDashboardSettings
                     }
                     return $returnVal
+                }
+                Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                    if ($Value -eq "Location")
+                    {
+                        return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                    }
+                    elseif ($Value -eq "dsn")
+                    {
+                        return "Connection string example"
+                    }
+                }
+                Mock -CommandName Test-Path {
+                    if ($Path -eq "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16\CONFIG\SPFlightRawConfig.json")
+                    {
+                        return $true
+                    }
+                    else
+                    {
+                        return $true
+                    }
+                }
+                if ($Global:SPDscHelper.CurrentStubBuildNumber.Major -eq 16 -and
+                    $Global:SPDscHelper.CurrentStubBuildNumber.Build -gt 13000)
+                {
+                    Mock -CommandName Update-SPFlightsConfigFile -MockWith { }
                 }
 
                 function Add-SPDscEvent
@@ -231,7 +256,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -281,7 +315,16 @@ try
                         RunCentralAdmin          = $false
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -330,7 +373,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -381,7 +433,16 @@ try
                         RunCentralAdmin          = $false
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -475,7 +536,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -558,7 +628,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         if ($global:SPDscConfigLockTriggered)
@@ -621,10 +700,6 @@ try
                         AdminContentDatabaseName = "SP_AdminContent"
                         RunCentralAdmin          = $true
                         CentralAdministrationUrl = "https://admin.contoso.com"
-                    }
-
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
                     }
 
                     Mock -CommandName Get-SPFarm -MockWith {
@@ -738,10 +813,6 @@ try
                         CentralAdministrationUrl = "http://localhost:8080"
                     }
 
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
-                    }
-
                     Mock -CommandName Get-SPFarm -MockWith {
                         return @{
                             Name                     = $testParams.FarmConfigDatabaseName
@@ -841,10 +912,6 @@ try
                         CentralAdministrationPort = 8080
                     }
 
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
-                    }
-
                     Mock -CommandName Get-SPFarm -MockWith {
                         return @{
                             Name                     = $testParams.FarmConfigDatabaseName
@@ -932,10 +999,6 @@ try
                         CentralAdministrationUrl  = "http://admin.contoso.com"
                         CentralAdministrationPort = 80
                         CentralAdministrationAuth = "Kerberos"
-                    }
-
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
                     }
 
                     Mock -CommandName Get-SPFarm -MockWith {
@@ -1075,10 +1138,6 @@ try
                         CentralAdministrationAuth = "NTLM"
                     }
 
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
-                    }
-
                     Mock -CommandName Get-SPFarm -MockWith {
                         return @{
                             Name                     = $testParams.FarmConfigDatabaseName
@@ -1213,10 +1272,6 @@ try
                         RunCentralAdmin           = $true
                         CentralAdministrationUrl  = "http://admin.contoso.com"
                         CentralAdministrationPort = 80
-                    }
-
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
                     }
 
                     Mock -CommandName Get-SPFarm -MockWith {
@@ -1358,10 +1413,6 @@ try
                         CentralAdministrationPort = 443
                     }
 
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
-                    }
-
                     Mock -CommandName Get-SPFarm -MockWith {
                         return @{
                             Name                     = $testParams.FarmConfigDatabaseName
@@ -1500,7 +1551,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -1578,7 +1638,16 @@ try
                         RunCentralAdmin          = $true
                     }
 
-                    Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                        if ($Value -eq "Location")
+                        {
+                            return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                        }
+                        elseif ($Value -eq "dsn")
+                        {
+                            return $null
+                        }
+                    }
                     Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                     Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                         return @{
@@ -1655,10 +1724,6 @@ try
                         RunCentralAdmin           = $true
                         CentralAdministrationUrl  = "https://admin.contoso.com"
                         CentralAdministrationPort = 443
-                    }
-
-                    Mock -CommandName Get-SPDscRegistryKey -MockWith {
-                        return "Connection string example"
                     }
 
                     Mock -CommandName Get-SPFarm -MockWith {
@@ -2265,7 +2330,16 @@ try
                             RunCentralAdmin          = $true
                         }
 
-                        Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                        Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                            if ($Value -eq "Location")
+                            {
+                                return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                            }
+                            elseif ($Value -eq "dsn")
+                            {
+                                return $null
+                            }
+                        }
                         Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                         Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                             return @{
@@ -2413,7 +2487,16 @@ try
                             RunCentralAdmin          = $false
                         }
 
-                        Mock -CommandName "Get-SPDscRegistryKey" -MockWith { return $null }
+                        Mock -CommandName Get-SPDscRegistryKey -MockWith {
+                            if ($Value -eq "Location")
+                            {
+                                return "C:\Program Files\Common Files\microsoft shared\Web Server Extensions\16"
+                            }
+                            elseif ($Value -eq "dsn")
+                            {
+                                return $null
+                            }
+                        }
                         Mock -CommandName "Get-SPFarm" -MockWith { return $null }
                         Mock -CommandName "Get-SPDscConfigDBStatus" -MockWith {
                             return @{