+++ CHANGELOG.md
@@ -5,6 +5,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- SPFarm
+  - Updated to run cmdlet `Update-SPFlightsConfigFile` on SharePoint Subscription.
+
 ## [5.4.0] - 2023-04-04
 
 ### Fixed